feat(finance): extract invoices list into feature components (Phase 6c)
- Extract table JSX into dedicated InvoiceRow component in InvoicesFeature
- Standardize InvoicesFeature toolbar using CardHeader with inline filters
- Enforce keyboard accessibility (role="button", onKeyDown) for invoice rows
