+++ features/invoices/index.ts
@@ -1,1 +1,2 @@
 export { InvoicesFeature } from './components/invoices-feature';
+export { InvoiceRow } from './components/invoice-row';
